Bases referencia/Emccamp/requirements.txt
pandas>=1.5
pyodbc>=4.0.35
pyyaml>=6.0
python-dotenv>=1.0
openpyxl>=3.1.0
requests>=2.31
